Adding callout to upstream change
The third step in the readme for generating the image includes special adjustment of the Dockerfile. This adjustment been merged in to the core dev containers repo: https://github.com/microsoft/vscode-dev-containers/pull/321

The instruction will be removed entirely once this behavior makes its way out to VS Code. For now, a callout in the instruction will suffice.